+++ src/braket/devices/local_simulator.py
@@ -286,8 +286,9 @@ class LocalSimulator(Device):
         return program
 
     @_construct_payload.register
-    def _(self, program: SerializableProgram, _inputs: dict[str, float], _shots: int):
-        return OpenQASMProgram(source=program.to_ir(ir_type=IRType.OPENQASM))
+    def _(self, program: SerializableProgram, inputs: Optional[dict[str, float]], _shots):
+        inputs_copy = inputs.copy() if inputs is not None else {}
+        return OpenQASMProgram(source=program.to_ir(ir_type=IRType.OPENQASM), inputs=inputs_copy)
 
     @_construct_payload.register
     def _(self, program: AnalogHamiltonianSimulation, _inputs: dict[str, float], _shots: int):